custom-chatbot-backend/requirements.txt
fastapi== 0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
python-socketio==5.11.3
transformers==4.42.3  # https://github.com/huggingface/transformers
sentence-transformers==3.0.1   # https://github.com/UKPLab/sentence-transformers
google-generativeai==0.7.2
python-dotenv==1.0.1
boto3==1.34.156
psycopg2==2.9.9
pyPDF2==3.0.1
firebase-admin==6.5.0